requirements.txt
asgiref==3.11.1
attrs==25.4.0
black==26.3.1
click==8.3.1
colorama==0.4.6
Django==6.0.3
django-filter==25.2
djangorestframework==3.16.1
drf-spectacular==0.29.0
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
mypy_extensions==1.1.0
packaging==26.0
pathspec==1.0.4
platformdirs==4.9.4
psycopg2==2.9.11
python-dotenv==1.2.2
pytokens==0.4.1
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
ruff==0.15.5
sqlparse==0.5.5
tzdata==2025.3
uritemplate==4.2.0
